Substitute the name of the identity consistently
The name of the identity-object is configurable. By default, it is just
Identity. If you pass an option to the generator, you can change it, to
say, Credential.

Failing to substitute this configurable name will cause references to
classes that don’t exist. If the identity-object is Credential, code
that refers to `Identity::CONST` will blow up.

To avoid code that says things like `identity = Credential.find(…)`,
make extra sure to substitute the correct name everywhere.